update factorial to double

+++ src/main/scala/com/akkineni/learn/Factorial.scala
@@ -3,17 +3,17 @@ package com.akkineni.learn
 import scala.annotation.tailrec
 
 object Factorial {
-  def compute(n: Int): Int = {
+  def compute(n: Int): Double = {
 
     @tailrec
-    def loop(index: Int, accum: Int): Int = {
+    def loop(index: Int, accum: Double): Double = {
       if (index == 1)
         accum
       else
         loop(index - 1, index * accum)
     }
 
-    loop(n, 1)
+    loop(n, 1.0)
   }
 
 }
